FUSED/requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py39h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cudatoolkit=11.3.1=h2bc3f7f_2
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=h4a9f257_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py39h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
pillow=10.2.0=py39h5eee18b_0
pip=23.3.1=py39h06a4308_0
python=3.9.19=h955ad1f_0
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_1
setuptools=68.2.2=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tmux=3.3a=h5eee18b_1
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
typing_extensions=4.9.0=py39h06a4308_1
urllib3=2.1.0=py39h06a4308_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
absl-py==2.1.0
accelerate==0.29.1
addict==2.4.0
arrow==1.3.0
asttokens==2.4.1
beautifulsoup4==4.12.3
cachetools==5.3.3
clip==0.2.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
flake8==3.7.9
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
gdown==5.1.0
grpcio==1.62.1
huggingface-hub==0.22.2
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.18.1
isort==4.3.21
jedi==0.19.1
joblib==1.3.2
jupyter-client==8.6.1
jupyter-core==5.7.2
kiwisolver==1.4.5
littleutils==0.2.2
markdown==3.6
markupsafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.6.1
nest-asyncio==1.6.0
netcal==1.3.5
nvidia-ml-py==12.535.133
nvitop==1.3.2
ogb==1.3.6
opencv-python==4.9.0.80
outdated==0.2.2
packaging==24.0
pandas==2.2.1
panopticapi==0.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0
pycodestyle==2.5.0
pyflakes==2.1.1
pygments==2.17.2
pyparsing==3.1.2
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.0.2
regex==2023.12.25
relplot==1.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
submitit==1.5.1
tabulate==0.9.0
tb-nightly==2.17.0a20240407
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.4.0
timm==0.9.16
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
wcwidth==0.2.13
werkzeug==3.0.2
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zipp==3.18.1
